+++ engine/runtime/src/main/scala/org/enso/interpreter/instrument/Handler.scala
@@ -4,7 +4,10 @@ import java.nio.ByteBuffer
 
 import com.oracle.truffle.api.TruffleContext
 import org.enso.interpreter.instrument.command.CommandFactory
-import org.enso.interpreter.instrument.execution.CommandProcessor
+import org.enso.interpreter.instrument.execution.{
+  EnsoCommandProcessor,
+  RuntimeContext
+}
 import org.enso.interpreter.service.ExecutionService
 import org.enso.polyglot.runtime.Runtime.Api
 import org.graalvm.polyglot.io.MessageEndpoint
@@ -56,9 +59,9 @@ final class Handler {
   val contextManager = new ExecutionContextManager
   val cache          = new Cache
 
-  var executionService: ExecutionService = _
-  var truffleContext: TruffleContext     = _
-  var commandProcessor: CommandProcessor = _
+  var executionService: ExecutionService     = _
+  var truffleContext: TruffleContext         = _
+  var commandProcessor: EnsoCommandProcessor = _
 
   /**
     * Initializes the handler with relevant Truffle objects, allowing it to
@@ -74,7 +77,7 @@ final class Handler {
     executionService = service
     truffleContext   = context
     endpoint.sendToClient(Api.Response(Api.InitializedNotification()))
-    commandProcessor = new CommandProcessor(1, executionService.getContext)
+    commandProcessor = new EnsoCommandProcessor(1, executionService.getContext)
   }
 
   /**
