+++ engine/runtime/src/main/scala/org/enso/interpreter/instrument/command/EditFileCmd.scala
@@ -1,6 +1,6 @@
 package org.enso.interpreter.instrument.command
 
-import org.enso.interpreter.instrument.RuntimeContext
+import org.enso.interpreter.instrument.execution.RuntimeContext
 import org.enso.polyglot.runtime.Runtime.Api
 
 import scala.jdk.CollectionConverters._
