+++ engine/runtime/src/main/scala/org/enso/interpreter/instrument/execution/RuntimeContext.scala
@@ -1,6 +1,11 @@
-package org.enso.interpreter.instrument
+package org.enso.interpreter.instrument.execution
 
 import com.oracle.truffle.api.TruffleContext
+import org.enso.interpreter.instrument.{
+  Cache,
+  Endpoint,
+  ExecutionContextManager
+}
 import org.enso.interpreter.service.ExecutionService
 
 case class RuntimeContext(
